+++ GHEtool/VariableClasses/PipeData/__init__.py
@@ -5,4 +5,5 @@ from .MultipleUTube import MultipleUTube
 from .CoaxialPipe import CoaxialPipe
 from .Separatus import Separatus
 from .Turbocollector import Turbocollector
+from .ConicalPipe import ConicalPipe
 from .PressureDrop import PressureDrop